parameters

all of the parameters in the EDCHM

Arguments

param_atmos_thr_Ts

<-1, 3> (Cel) threshold air temperature that snow, parameter for atmosSnow_ThresholdT()

param_atmos_ubc_A0FORM

<0.01, 3> (Cel) threshold air temperature that snow, it can not equal or small than 0, parameter for atmosSnow_UBC()

param_baseflow_grf_gamma

<2, 7> exponential parameter for baseflow_GR4Jfix()

param_baseflow_sur_k

<0.01, 1> coefficient parameter for baseflow_SupplyRatio()

param_baseflow_sup_k

<0.01, 1> coefficient parameter for baseflow_SupplyPow()

param_baseflow_sup_gamma

<0, 1> exponential parameter for baseflow_SupplyPow()

param_baseflow_map_gamma

<0.1, 5> exponential parameter for baseflow_MaxPow()

param_baseflow_thp_thresh

<0.1, 0.9> coefficient parameter for baseflow_ThreshPow()

param_baseflow_thp_gamma

<0.1, 5> exponential parameter for baseflow_ThreshPow()

param_baseflow_arn_thresh

<0.1, 0.9> coefficient parameter for baseflow_ThreshPow()

param_baseflow_arn_k

<0.1, 1> exponential parameter for baseflow_ThreshPow()

param_capirise_acr_k

<0.01, 1> coefficient parameter capirise_AcceptRatio()

param_capirise_acp_k

<0.01, 1> coefficient parameter for capirise_AcceptPow()

param_capirise_acp_gamma

<0.01, 1> exponential parameter for capirise_AcceptPow()

param_confluen_kel_k

<1, 4> parameter forconfluenIUH_Kelly()

param_confluen_nas_n

<1, 8> parameter forconfluenIUH_Nash()

param_evatrans_tur_k

<0.6, 1> parameter for evatransPotential_TurcWendling(), higher value when closer to the sea

param_evatrans_sur_k

<0.1, 1> parameter for evatransActual_SupplyRatio(), ratio of potential ET, that is estimated as actually ET

param_evatrans_sup_k

<0.1, 1> parameter for evatransActual_SupplyPow(), ratio of this method

param_evatrans_sup_gamma

<1, 5> parameter for evatransActual_SupplyPow(), exponent of this method

param_evatrans_vic_gamma

<0.2, 5> parameter for evatransActual_VIC()

param_evatrans_ubc_gamma

<0.5, 2> parameter for evatransActual_UBC()

param_evatrans_lia_gamma

<0.4, 1> parameter for evatransActual_LiangLand()

param_evatrans_lia_B

<0.01, 3> parameter for evatransActual_LiangSoil()

param_infilt_ubc_P0AGEN

<0.1, 4> coefficient parameter for infilt_UBC()

param_infilt_sur_k

<0.01, 1> coefficient parameter for infilt_SupplyRatio()

param_infilt_acr_k

<0.01, 1> coefficient parameter for infilt_AcceptRatio()

param_infilt_sup_k

<0.01, 1> coefficient parameter for infilt_SupplyPow()

param_infilt_sup_gamma

<0, 1> parameters for infilt_SupplyPow()

param_infilt_acp_k

<0.01, 1> coefficient parameter for infilt_AcceptPow()

param_infilt_acp_gamma

<0.001, 5> parameters for infilt_AcceptPow()

param_infilt_hbv_beta

<0.001, 5> parameters for infilt_HBV()

param_infilt_xaj_B

<0.01, 3> parameters for infilt_XAJ()

param_infilt_vic_B

<0.01, 3> parameters for infilt_VIC()

param_inteflow_grf_k

<0.01, 1> coefficient parameter for inteflow_GR4Jfix()

param_inteflow_grf_gamma

<2, 7> exponential parameter for baseflow_GR4Jfix()

param_inteflow_map_gamma

<0.1, 5> exponential parameter for inteflow_MaxPow()

param_inteflow_thp_thresh

<0.1, 0.9> coefficient parameter for inteflow_ThreshPow()

param_inteflow_thp_gamma

<0.1, 5> exponential parameter for inteflow_ThreshPow()

param_inteflow_arn_thresh

<0.1, 0.9> coefficient parameter for inteflow_ThreshPow()

param_inteflow_arn_k

<0.1, 1> exponential parameter for inteflow_ThreshPow()

param_inteflow_sp0_k

<0.01, 1> coefficient parameter for inteflow_SupplyPow0()

param_inteflow_sp0_gamma

<0, 1> exponential parameter for inteflow_SupplyPow0()

param_inteflow_sup_k

<0.01, 1> coefficient parameter for inteflow_SupplyPow()

param_inteflow_sup_gamma

<0, 7> parameter for inteflow_SupplyPow()

param_inteflow_sur_k

<0.01, 1> coefficient parameter for inteflow_SupplyRatio()

param_lateral_sup_k

<-1, 1> coefficient parameter for lateral_SupplyPow()

param_lateral_sup_gamma

<0.01, 5> parameters for lateral_SupplyPow()

param_lateral_sur_k

<-2, 1> coefficient parameter for lateral_SupplyRatio()

param_lateral_grf_gamma

<0.01, 5> parameter for lateral_GR4Jfix()

param_lateral_thp_thresh

<0.1, 0.9> coefficient parameter for lateral_ThreshPow()

param_lateral_thp_gamma

<0.1, 5> exponential parameter for lateral_ThreshPow()

param_lateral_arn_thresh

<0.1, 0.9> coefficient parameter for lateral_ThreshPow()

param_lateral_arn_k

<0.1, 1> exponential parameter for lateral_ThreshPow()

param_percola_grf_k

<0.01, 1> coefficient parameter for percola_GR4Jfix()

param_percola_map_gamma

<0.1, 5> exponential parameter for percola_MaxPow()

param_percola_thp_thresh

<0.1, 0.9> coefficient parameter for percola_ThreshPow()

param_percola_thp_gamma

<0.1, 5> exponential parameter for percola_ThreshPow()

param_percola_arn_thresh

<0.1, 0.9> coefficient parameter for percola_ThreshPow()

param_percola_arn_k

<0.1, 1> exponential parameter for percola_ThreshPow()

param_percola_sup_k

<0.01, 1> coefficient parameter for percola_SupplyPow()

param_percola_sup_gamma

<0, 7> parameter for percola_SupplyPow()

param_percola_sur_k

<0.01, 1> coefficient parameter for percola_SupplyRatio()

param_snow_kus_fE

<0.0005, 0.003> (mm/m2/MJ) snow melt temperature parameter for snowMelt_Factor()

param_snow_kus_fT

<0.05, 1> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for snowMelt_Factor()

param_snow_fac_Tmelt

<0, 3> (Cel) snow melt temperature parameter for snowMelt_Factor()

param_snow_fac_f

<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for snowMelt_Factor()