functions for atmosphere

calculate some basic physical, meteorological variables

atmos_NettoRadiat(
  time_dayOfYear_,
  atmos_temperature_Cel,
  atmos_temperatureMax_Cel,
  atmos_temperatureMin_Cel,
  atmos_relativeHumidity_1,
  atmos_solarRadiat_MJ,
  land_latitude_Degree,
  land_elevation_m
)

atmos_SaturatVaporPress(atmos_temperature_Cel)

atmos_VaporPress(atmos_temperature_Cel, atmos_relativeHumidity_1)

atmos_WindSpeed2m(atmos_windSpeed_m_s, atmos_windMeasureHeight_m)

Arguments

time_dayOfYear_

<1, 366> (d) the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)

atmos_temperature_Cel

(Cel) the average air temperature in the time phase

atmos_temperatureMax_Cel

(Cel) the maximal air temperature in the time phase

atmos_temperatureMin_Cel

(Cel) the minimal air temperature in the time phase

atmos_relativeHumidity_1

(0, 1) relative humidity

atmos_solarRadiat_MJ

(MJ/m2/TS) the solar radiation that actually reaches the earths surface

land_latitude_Degree

(degree) average latitude

land_elevation_m

(m) average elevation

atmos_windSpeed_m_s

(m/s) measured wind speed at z m above ground surface

atmos_windMeasureHeight_m

(m) height of measurement above ground surface

Value

atmos_netRadiat_MJ (MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earths surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation

atmos_netRadiat_MJ (MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earth's surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation

atmos_vaporPress_hPa (hPa) actual vapour pressure, can be calculated by atmos_VaporPress()

atmos_WindSpeed2m_m_s (m/s) wind speed at 2 m above ground surface