potential evapotranspiration

The concept of potential evapotranspiration (ET) estimates the ability of water lost from the soil and vegetation in an area due to evaporation and transpiration. It assumes that there is always enough water in the ET area to meet the demand for evapotranspiration. However, the characteristics of the ET area, such as whether it is covered with vegetation or bare soil, can affect the amount of evapotranspiration that occurs. In order to accurately estimate potential ET, we need to consider these characteristics.

But we may not always have access to the necessary information or effective methods to do this. In these cases, we can use a simplified method known as reference ET. This method defines the ET area using certain fixed characteristics, such as those provided by the evatransPotential_FAO56() function. In this situation, we need to provide factors to account for the differences between the actual ET area and the reference ET area.

evatransPotential_TurcWendling(
  atmos_temperature_Cel,
  atmos_solarRadiat_MJ,
  time_step_h,
  param_evatrans_tur_k
)

evatransPotential_Linacre(
  time_dayOfYear_,
  atmos_temperature_Cel,
  atmos_relativeHumidity_1,
  land_latitude_Degree,
  land_elevation_m,
  land_albedo_1
)

evatransPotential_FAO56(
  time_dayOfYear_,
  atmos_temperature_Cel,
  atmos_vaporPress_hPa,
  atmos_saturatVaporPress_hPa,
  atmos_netRadiat_MJ,
  atmos_windSpeed2m_m_s,
  land_latitude_Degree,
  land_elevation_m
)

Arguments

atmos_temperature_Cel

(Cel) the average air temperature in the time phase

atmos_solarRadiat_MJ

(MJ/m2/TS) the solar radiation that actually reaches the earths surface

time_step_h

<1, 24> (h) time step in hour

param_evatrans_tur_k

<0.6, 1> parameter for evatransPotential_TurcWendling(), higher value when closer to the sea

time_dayOfYear_

<1, 366> (d) the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)

atmos_relativeHumidity_1

(0, 1) relative humidity

land_latitude_Degree

(degree) average latitude

land_elevation_m

(m) average elevation

land_albedo_1

<0, 1> albedo of the region

atmos_vaporPress_hPa

(hPa) actual vapour pressure, can be calculated by atmos_VaporPress()

atmos_saturatVaporPress_hPa

(hPa) saturation vapour pressure at atmos_temperature_Cel, can be calculated by atmos_SaturatVaporPress()

atmos_netRadiat_MJ

(MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earths surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation

atmos_windSpeed2m_m_s

(m/s) wind speed at 2 m above ground surface

Value

potential evapotranspiration (mm/m2)

Details

_TurcWendling (WENDLING et al. 1991) : consider only the radiation and temperature as the main factors. \[E_p = \frac{(100 R_s + 3.875 t_h k)\cdot(T + 22)}{150 (T + 123)}\] where

\(E_p\) is potential ET, atmos_potentialEvatrans_mm

\(R_s\) is solar radiation, atmos_solarRadiat_MJ

\(t_h\) is time step in hour, time_step_h

\(T\) is average air temperature, atmos_temperature_Cel

\(k\) is param_evatrans_tur_k

_Linacre (Linacre 1977) : consider only the temperature as the main factors. \[E_p = \frac{\frac{100(0.75 - \alpha)(T + 0.006 z)}{100 - \phi} + 15(T - T_d)}{80 - T}\] \[T_d = T - 20 (1-H_R)\] where

\(\alpha\) is albedo, land_albedo_1

\(z\) is elevation, land_elevation_m

\(T_d\) is dewpoint temperature,

\(H_R\) is relative humidity, atmos_relativeHumidity_1

_FAO56 (Allen et al. 1998) : consider not only radiation and temperature but also other variable like wind speed as the main factors. \[E_p =\frac{0.408 \Delta\left(R_n - G\right)+\gamma \frac{900}{T+273} {u}_{2}\left({e}_{{s}}-{e}_{{a}}\right)}{\Delta+\gamma\left(1+0.34 {u}_{2}\right)}\] where

\(\Delta\) is slope vapour pressure curve (kPa °C-1)

\(R_n\) is net radiation, atmos_netRadiat_MJ

\(G\) is soil heat flux density

\(u_2\) is wind speed at 2 m height, atmos_windSpeed2m_m_s

\(e_s\) is saturation vapour pressure, atmos_saturatVaporPress_hPa

\(e_a\) is actual vapour pressure, atmos_vaporPress_hPa

\(\gamma\) is psychrometric constant

References

Allen RG, Pereira LS, Raes D, Smith M (1998). “Crop Evapotranspiration. Guidelines for Computing Crop Water Requirements.” FAO Irrigation and Drainage Paper (FAO).

Linacre ET (1977). “A Simple Formula for Estimating Evaporation Rates in Various Climates, Using Temperature Data Alone.” Agricultural Meteorology, 18(6), 409--424. ISSN 0002-1571, doi: 10.1016/0002-1571(77)90007-3 .

WENDLING U, SCHELLIN H, THOM挼㸴 M (1991). “Bereitstellung von T攼㸴glichen Informationen Zum Wasserhaushalt Des Bodens F昼㹣r Die Zwecke Der Agrarmeteorologischen Beratung.” Bereitstellung von t攼㸴glichen Informationen zum Wasserhaushalt des Bodens f昼㹣r die Zwecke der agrarmeteorologischen Beratung, 41(6), 468--475. ISSN 0084-5361.