variable define

In this topic all of the variable in the EDCHM will be defined.

Arguments

time_dayOfYear_

<1, 366> (d) the number of the day in the year between 1 (1 January) and 365 or 366 (31 December)

atmos_precipitation_mm

(mm/m2/TS) precipitaion volum

atmos_rain_mm

(mm/m2/TS) precipitation in rain form

atmos_snow_mm

(mm/m2/TS) precipitation in snow form

atmos_temperature_Cel

(Cel) the average air temperature in the time phase

atmos_temperatureMax_Cel

(Cel) the maximal air temperature in the time phase

atmos_temperatureMin_Cel

(Cel) the minimal air temperature in the time phase

atmos_solarRadiat_MJ

(MJ/m2/TS) the solar radiation that actually reaches the earths surface

atmos_netRadiat_MJ

(MJ/m2/TS) the balance between the energy absorbed, reflected and emitted by the earths surface or the difference between the incoming net shortwave (Rns) and the net outgoing longwave (Rnl) radiation

atmos_relativeHumidity_1

(0, 1) relative humidity

atmos_vaporPress_hPa

(hPa) actual vapour pressure, can be calculated by atmos_VaporPress()

atmos_saturatVaporPress_hPa

(hPa) saturation vapour pressure at atmos_temperature_Cel, can be calculated by atmos_SaturatVaporPress()

atmos_windSpeed_m_s

(m/s) measured wind speed at z m above ground surface

atmos_windMeasureHeight_m

(m) height of measurement above ground surface

atmos_windSpeed2m_m_s

(m/s) wind speed at 2 m above ground surface

atmos_potentialEvatrans_mm

(mm/m2/TS) potential / reference evapotranspiration

land_albedo_1

<0, 1> albedo of the region

land_latitude_Degree

(degree) average latitude

land_elevation_m

(m) average elevation

land_impermeableFrac_1

<0, 1> the maximum impermeable fraction when th soil is fully saturated

land_water_mm

(mm/m2) pounded water volume in landLy and there is no limit, different than land_interceptWater_mm

land_interceptWater_mm

(mm/m2) initial water volume that can be intercepted

land_interceptCapacity_mm

(mm/m2) average intercept Capacity (maximal storage capacity)

land_actualEvatrans_mm

(mm/m2/TS) actual evapotranspiration from landLy

land_infiltrat_mm

(mm/m2/TS) infiltration from landLy to soilLy

snow_ice_mm

(mm/m2) water equivalent of ice in snowpack

soil_fieldCapacityPerc_1

<0, 1> the relative ratio (\(\theta_fc / \theta^*\)) that the water content can drainage by gravity

soil_water_mm

(mm/m2) water volume in soilLy

soil_capacity_mm

(mm/m2) average soil Capacity (maximal storage capacity)

soil_interflow_mm

(mm/m2/TS) subsurface flow directly to the river

soil_actualEvatrans_mm

(mm/m2/TS) actual evapotranspiration from soilLy

soil_potentialPercola_mm

<0.01, 7> (mm/m2/TS) potential percolation

soil_potentialInteflow_mm

<0.01, 7> (mm/m2/TS) potential interflow

soil_potentialCapirise_mm

<0.01, 7> (mm/m2/TS) potential capillary rise

ground_water_mm

(mm/m2/TS) water volume in groundLy

ground_capacity_mm

(mm/m2) water storage capacity in groundLy

ground_lateral_mm

(mm/m2/TS) lateral flow, exchange with outside region. It can be NEGATIV

ground_capillarise_mm

(mm/m2/TS) capillary rise from groundLy to soilLy

ground_potentialLateral_mm

<-7, 7> (mm/m2/TS) potential lateral flow

ground_potentialBaseflow_mm

<0.01, 7> (mm/m2/TS) potential baseflow

confluen_inputWater_mm, land_runoff_mm, ground_baseflow_mm

(mm/m2) input water volum in every routeline

confluen_iuh_1, confluen_iuhLand_1, confluen_iuhSoil_1, confluen_iuhGround_1

(vector of num, sume() = 1) the ratio in every timestep, can be calculated by confluenIUH_GR4J1(), confluenIUH_GR4J2()

confluen_responseTime_TS

(TS) response or concentration time in every routeline

water_mm

(mm/m2/TS) water volume in soilLy, groundLy or intercept of landLy, same as soil_water_mm, ground_water_mm or land_interceptWater_mm

capacity_mm

(mm/m2) water storage capacity in soilLy, groundLy or intercept of landLy, same as soil_capacity_mm, ground_capacity_mm orland_interceptCapacity_mm