confluence

In hydrological modeling, routing (named as confluen in EDCHM) refers to the process of simulating the movement of water through a river network or other drainage system. It allows the model to predict the flow of water in rivers and streams. In hydrological models, routing is typically performed using mathematical algorithms that account for the physical properties of the river network, such as its geometry, roughness, and discharge capacity. The parameters that govern routing, such as flow velocity and channel roughness, can have a significant impact on the accuracy of the model.

confluence is a calculation function that causes water to flow into the gauge point.

IUH: IUH (Instant Unit Hydrograph) with one watercourse,

IUH2S: IUH with two water sources, each with a different IUH vector,

IUH3S: IUH with three water sources, each with a different IUH vector.

Under the concept of the conceptual HM, the water flux to the water flow will be calculated using the confluence process. This process does not calculate the water balance, but rather the time-varying nature of the water flow. The "Instant Unit Hydrograph" method is the most effective way to deal with time-varying flows. In the first stage, only confluenIUH will be supported.

So we can give the function:

\[Q = f_{confluen}(F, u)\]

where

\(Q\) is stream flow, but still in mm/TS not m3/TS or m3/S

\(F\) is flux that will into river conflen, e.g.land_runoff_mm, soil_interflow_mm or ground_baseflow_mm

\(u\) is Instant Unit Hydrograph series

confluen_IUH(confluen_inputWater_mm, confluen_iuh_1)

confluen_IUH2S(
  land_runoff_mm,
  ground_baseflow_mm,
  confluen_iuhLand_1,
  confluen_iuhGround_1
)

confluen_IUH3S(
  land_runoff_mm,
  soil_interflow_mm,
  ground_baseflow_mm,
  confluen_iuhLand_1,
  confluen_iuhSoil_1,
  confluen_iuhGround_1
)

Arguments

confluen_inputWater_mm, land_runoff_mm, ground_baseflow_mm

(mm/m2) input water volum in every routeline

confluen_iuh_1, confluen_iuhLand_1, confluen_iuhSoil_1, confluen_iuhGround_1

(vector of num, sume() = 1) the ratio in every timestep, can be calculated by confluenIUH_GR4J1(), confluenIUH_GR4J2()

soil_interflow_mm

(mm/m2/TS) subsurface flow directly to the river

Value

confluenced water (mm/m2)

References

There are no references for Rd macro \insertAllCites on this help page.