modells build with EDCHM modulas

some example models with EDCHM modulas

some example models with EDCHM modulas

Usage

EDCHM_GR4J(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  S_,
  R_,
  X_1,
  X_2,
  X_3,
  X_4
)

EDCHM_GR4J_full(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  S_,
  R_,
  X_1,
  X_2,
  X_3,
  X_4
)

EDCHM_mini(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh
)

EDCHM_mini_full(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh
)

EDCHM_snow(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  atmos_temperature_Cel,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  snow_ice_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_atmos_thr_Ts,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh,
  param_snow_fac_f,
  param_snow_fac_Tmelt
)

EDCHM_snow_full(
  n_time,
  n_spat,
  atmos_potentialEvatrans_mm,
  atmos_precipitation_mm,
  atmos_temperature_Cel,
  ground_capacity_mm,
  ground_water_mm,
  land_impermeableFrac_1,
  snow_ice_mm,
  soil_capacity_mm,
  soil_potentialPercola_mm,
  soil_water_mm,
  confluenLand_responseTime_TS,
  confluenGround_responseTime_TS,
  param_atmos_thr_Ts,
  param_baseflow_grf_gamma,
  param_confluenLand_kel_k,
  param_evatrans_ubc_gamma,
  param_infilt_ubc_P0AGEN,
  param_percola_arn_k,
  param_percola_arn_thresh,
  param_snow_fac_f,
  param_snow_fac_Tmelt
)

Arguments

n_time, n_spat

number of time step and spatial unit

atmos_potentialEvatrans_mm

(mm/m2/TS) potential / reference evapotranspiration

atmos_precipitation_mm

(mm/m2/TS) precipitaion volum

S_, R_

storage water S and R

X_1, X_2, X_3, X_4

parameters in GR4J

ground_capacity_mm

(mm/m2) water storage capacity in groundLy

ground_water_mm

(mm/m2/TS) water volume in groundLy

land_impermeableFrac_1

<0, 1> the maximum impermeable fraction when th soil is fully saturated

soil_capacity_mm

(mm/m2) average soil Capacity (maximal storage capacity)

soil_potentialPercola_mm

<0.01, 7> (mm/m2/TS) potential percolation

soil_water_mm

(mm/m2) water volume in soilLy

confluenLand_responseTime_TS, confluenGround_responseTime_TS, param_confluenLand_kel_k,

parameters in modells

param_baseflow_grf_gamma

<2, 7> exponential parameter for baseflow_GR4Jfix()

param_evatrans_ubc_gamma

<0.5, 2> parameter for evatransActual_UBC()

param_infilt_ubc_P0AGEN

<0.1, 4> coefficient parameter for infilt_UBC()

param_percola_arn_k

<0.1, 1> exponential parameter for percola_ThreshPow()

param_percola_arn_thresh

<0.1, 0.9> coefficient parameter for percola_ThreshPow()

atmos_temperature_Cel

(Cel) the average air temperature in the time phase

snow_ice_mm

(mm/m2) water equivalent of ice in snowpack

param_atmos_thr_Ts

<-1, 3> (Cel) threshold air temperature that snow, parameter for atmosSnow_ThresholdT()

param_snow_fac_f

<0.05, 2> (mm/m2/h/Cel) potential melt volum per Cel per hour parameter for snowMelt_Factor()

param_snow_fac_Tmelt

<0, 3> (Cel) snow melt temperature parameter for snowMelt_Factor()

Value

EDCHM_xxxx: stream flow in mm/TS

EDCHM_xxxx_full: list of variablen

stream flow in mm/TS

EDCHM_GR4J (Perrin et al. 2003) :

Total same like original GR4J

EDCHM_GR4J_full (Perrin et al. 2003) :

With all variable output

EDCHM_mini:

A model based on mini-structure with only six process:

evatransActual (evatransSoil)

infilt

percola

baseflow

confluenIUH (confluenLand)

confluenIUH (confluenGround)

EDCHM_mini_full:

With full Output

EDCHM_snow:

A model with snowMelt and atmosSnow based on mini-modell

EDCHM_snow_full:

With full Output

References

Perrin C, Michel C, Andr攼㸹assian V (2003). “Improvement of a Parsimonious Model for Streamflow Simulation.” Journal of Hydrology, 279(1-4), 275--289. ISSN 00221694, doi: 10.1016/S0022-1694(03)00225-7 .

Perrin C, Michel C, Andr攼㸹assian V (2003). “Improvement of a Parsimonious Model for Streamflow Simulation.” Journal of Hydrology, 279(1-4), 275--289. ISSN 00221694, doi: 10.1016/S0022-1694(03)00225-7 .